+++ web/src/compose_validate.ts
@@ -1075,6 +1075,28 @@ export function check_overflow_text($container: JQuery): number {
     return text.length;
 }
 
+export let check_posting_policy_for_compose_box = (): void => {
+    const banner_text = get_posting_policy_error_message();
+    if (banner_text === "") {
+        compose_banner.clear_errors();
+        return;
+    }
+
+    let banner_classname = compose_banner.CLASSNAMES.no_post_permissions;
+    if (compose_state.selected_recipient_id === "direct") {
+        banner_classname = compose_banner.CLASSNAMES.cannot_send_direct_message;
+        compose_banner.cannot_send_direct_message_error(banner_text);
+    } else {
+        compose_banner.show_error_message(banner_text, banner_classname, $("#compose_banners"));
+    }
+};
+
+export function rewire_check_posting_policy_for_compose_box(
+    value: typeof check_posting_policy_for_compose_box,
+): void {
+    check_posting_policy_for_compose_box = value;
+}
+
 export let validate_and_update_send_button_status = function (): void {
     const is_valid = validate(false, false);
     const $send_button = $("#compose-send-button");
@@ -1086,6 +1108,7 @@ export let validate_and_update_send_button_status = function (): void {
         send_button_element._tippy.hide();
         send_button_element._tippy.show();
     }
+    check_posting_policy_for_compose_box();
 };
 
 export function rewire_validate_and_update_send_button_status(
